♻️ refactor(agent-runtime): move empty-completion + reasoning-replay utils into the package
Sink transport-independent runtime helpers from the server executor layer into
@lobechat/agent-runtime so both client and server can share them (first slice of
LOBE-10949 P2 — "下沉 runtime-pure 逻辑进包"):

- ModelEmptyError + isEmptyModelCompletion -> core/modelCompletion.ts
- stripAssistantReasoningForReplay        -> utils/replay.ts

The server executor layer re-imports them from the package; the local
ModelEmptyError.ts shim is removed and all call sites (incl. tests) repointed.
Only helpers whose deps are @lobechat/types-only were moved — retry budgets and
LLM error classification stay server-side (they need model-runtime /
business-const) and are tracked as a follow-up.

Behavior unchanged: type-check clean, agent-runtime suite 266/266,
RuntimeExecutors.test.ts 125/125.

Part of LOBE-10949
