+++ packages/agent-runtime/src/core/index.ts
@@ -1,3 +1,4 @@
 export * from './InterventionChecker';
+export * from './modelCompletion';
 export * from './runtime';
 export * from './UsageCounter';
