+++ packages/agent-runtime/src/utils/index.ts
@@ -1,4 +1,5 @@
 export * from './messageSelectors';
+export * from './replay';
 export * from './status';
 export * from './stepContextComputer';
 export * from './tokenCounter';
